generic sort trait
