+++ src/flatbush/mod.rs
@@ -2,9 +2,11 @@ pub mod builder;
 pub mod constants;
 pub mod error;
 pub mod index;
+pub mod sort;
 pub mod r#trait;
 pub mod util;
 
 pub use builder::FlatbushBuilder;
 pub use index::{FlatbushRef, OwnedFlatbush};
 pub use r#trait::FlatbushIndex;
+pub use sort::HilbertSort;
