+++ src/flatbush/sort/mod.rs
@@ -1,0 +1,5 @@
+pub mod hilbert;
+pub mod r#trait;
+
+pub use hilbert::HilbertSort;
+pub use r#trait::{Sort, SortParams};
